Flask/requirements.txt
Pillow
torch
torchvision
opencv-python
imutils
Flask
ultralytics
flask_cors
